Bug 959597 part 3 - Add ARM simulator for JIT code. r=nbp

+++ js/src/jit/arm/BaselineIC-arm.cpp
@@ -71,7 +71,7 @@ ICCompare_Double::Compiler::generateStubCode(MacroAssembler &masm)
 // ICBinaryArith_Int32
 
 extern "C" {
-    extern int __aeabi_idivmod(int,int);
+    extern int64_t __aeabi_idivmod(int,int);
 }
 
 bool
