+++ js/src/jit/arm/CodeGenerator-arm.cpp
@@ -609,8 +609,8 @@ CodeGeneratorARM::visitDivI(LDivI *ins)
 }
 
 extern "C" {
-    extern int __aeabi_idivmod(int,int);
-    extern int __aeabi_uidivmod(int,int);
+    extern int64_t __aeabi_idivmod(int,int);
+    extern int64_t __aeabi_uidivmod(int,int);
 }
 
 bool
